attempted fix for dedicated servers
