+++ src/main/java/net/onixary/shapeShifterCurseFabric/form_giving_custom_entity/TEntitySpawnHandler.java
@@ -1,54 +1,28 @@
 package net.onixary.shapeShifterCurseFabric.form_giving_custom_entity;
 
-import net.fabricmc.fabric.api.client.rendering.v1.EntityModelLayerRegistry;
-import net.fabricmc.fabric.api.client.rendering.v1.EntityRendererRegistry;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerEntityEvents;
 import net.fabricmc.fabric.api.object.builder.v1.entity.FabricDefaultAttributeRegistry;
-import net.minecraft.client.render.entity.AxolotlEntityRenderer;
-import net.minecraft.client.render.entity.model.EntityModelLayer;
-import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.passive.AxolotlEntity;
 import net.minecraft.entity.passive.BatEntity;
 import net.minecraft.entity.passive.OcelotEntity;
-import net.minecraft.util.Identifier;
 import net.minecraft.world.World;
-import net.onixary.shapeShifterCurseFabric.ShapeShifterCurseFabric;
-import net.onixary.shapeShifterCurseFabric.form_giving_custom_entity.axolotl.TAxolotlEntityRenderer;
+import net.onixary.shapeShifterCurseFabric.client.ShapeShifterCurseFabricClient;
 import net.onixary.shapeShifterCurseFabric.form_giving_custom_entity.axolotl.TransformativeAxolotlEntity;
-import net.onixary.shapeShifterCurseFabric.form_giving_custom_entity.bat.BatEntityModel;
-import net.onixary.shapeShifterCurseFabric.form_giving_custom_entity.bat.BatEntityRenderer;
 import net.onixary.shapeShifterCurseFabric.form_giving_custom_entity.bat.TransformativeBatEntity;
-import net.onixary.shapeShifterCurseFabric.form_giving_custom_entity.ocelot.TOcelotEntityRenderer;
 import net.onixary.shapeShifterCurseFabric.form_giving_custom_entity.ocelot.TransformativeOcelotEntity;
 
 import static net.onixary.shapeShifterCurseFabric.ShapeShifterCurseFabric.*;
-//import static net.onixary.shapeShifterCurseFabric.data.StaticParams.T_AXOLOTL_REPLACE_PROBABILITY;
-//import static net.onixary.shapeShifterCurseFabric.data.StaticParams.T_BAT_REPLACE_PROBABILITY;
 
 public class TEntitySpawnHandler {
-    public static final EntityModelLayer T_BAT_LAYER = new EntityModelLayer(new Identifier(MOD_ID, "t_bat"), "main");
-    public static final EntityModelLayer T_AXOLOTL_LAYER = new EntityModelLayer(new Identifier(MOD_ID, "t_axolotl"), "main");
-    public static final EntityModelLayer T_OCELOT_LAYER = new EntityModelLayer(new Identifier(MOD_ID, "t_ocelot"), "main");
 
     public static void register() {
         // Reg custom entities model and renderer
         // bat
         FabricDefaultAttributeRegistry.register(T_BAT, TransformativeBatEntity.createTBatAttributes());
-        EntityRendererRegistry.register(T_BAT, (context) -> {
-            return new BatEntityRenderer(context);
-        });
-        EntityModelLayerRegistry.registerModelLayer(T_BAT_LAYER, BatEntityModel::getTexturedModelData);
         // axolotl
         FabricDefaultAttributeRegistry.register(T_AXOLOTL, TransformativeAxolotlEntity.createTAxolotlAttributes());
-        EntityRendererRegistry.register(T_AXOLOTL, (context) -> {
-            return new TAxolotlEntityRenderer(context);
-        });
-        EntityModelLayerRegistry.registerModelLayer(T_AXOLOTL_LAYER, BatEntityModel::getTexturedModelData);
         // ocelot
         FabricDefaultAttributeRegistry.register(T_OCELOT, TransformativeBatEntity.createTBatAttributes());
-        EntityRendererRegistry.register(T_OCELOT, (context) -> {
-            return new TOcelotEntityRenderer(context);
-        });
 
 
         // handle entity spawn
