+++ src/main/java/net/onixary/shapeShifterCurseFabric/item/BookOfShapeShifter.java
@@ -1,29 +1,22 @@
 package net.onixary.shapeShifterCurseFabric.item;
 
-import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.item.TooltipContext;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.server.network.ServerPlayerEntity;
-import net.minecraft.server.world.ServerWorld;
 import net.minecraft.text.Text;
 import net.minecraft.util.Formatting;
 import net.minecraft.util.Hand;
 import net.minecraft.util.TypedActionResult;
 import net.minecraft.world.World;
 import net.onixary.shapeShifterCurseFabric.ShapeShifterCurseFabric;
-import net.onixary.shapeShifterCurseFabric.custom_ui.BookOfShapeShifterScreen;
-import net.onixary.shapeShifterCurseFabric.custom_ui.StartBookScreen;
-import net.onixary.shapeShifterCurseFabric.integration.origins.origin.Origin;
-import net.onixary.shapeShifterCurseFabric.integration.origins.origin.OriginLayer;
+import net.onixary.shapeShifterCurseFabric.client.ShapeShifterCurseFabricClient;
 import net.onixary.shapeShifterCurseFabric.player_form.PlayerForms;
-import net.onixary.shapeShifterCurseFabric.player_form.ability.PlayerFormComponent;
 import net.onixary.shapeShifterCurseFabric.player_form.ability.RegPlayerFormComponent;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
-import java.util.Map;
 
 public class BookOfShapeShifter extends Item {
     public BookOfShapeShifter(Settings settings) {
@@ -35,20 +28,12 @@ public class BookOfShapeShifter extends Item {
         PlayerForms currentForm = user.getComponent(RegPlayerFormComponent.PLAYER_FORM).getCurrentForm();
         if (world.isClient) {
             // 客户端逻辑：仅处理打开界面
-            if (currentForm == PlayerForms.ORIGINAL_BEFORE_ENABLE) {
-                if (!(MinecraftClient.getInstance().currentScreen instanceof StartBookScreen)) {
-                    StartBookScreen startScreen = new StartBookScreen();
-                    startScreen.currentPlayer = user;
-                    MinecraftClient.getInstance().setScreen(startScreen);
-                }
-            } else if (!(MinecraftClient.getInstance().currentScreen instanceof BookOfShapeShifterScreen)) {
-                BookOfShapeShifterScreen bookScreen = new BookOfShapeShifterScreen();
-                bookScreen.currentPlayer = user;
-                MinecraftClient.getInstance().setScreen(bookScreen);
-            }
+            if (currentForm == PlayerForms.ORIGINAL_BEFORE_ENABLE)
+                ShapeShifterCurseFabricClient.openStartBookScreen(user);
+            else ShapeShifterCurseFabricClient.openBookScreen(user);
         } else {
             // 服务端逻辑：触发成就
-            if (currentForm != PlayerForms.ORIGINAL_BEFORE_ENABLE){
+            if (currentForm != PlayerForms.ORIGINAL_BEFORE_ENABLE) {
                 if (user instanceof ServerPlayerEntity serverPlayer) {
                     ShapeShifterCurseFabric.ON_OPEN_BOOK_OF_SHAPE_SHIFTER.trigger(serverPlayer);
                 }
