+++ src/main/resources/shape-shifter-curse.mixins.json
@@ -4,28 +4,13 @@
   "package": "net.onixary.shapeShifterCurseFabric.mixin",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
-    "OverrideSkinFirstPersonMixin",
-    "OverrideSkinThirdPersonMixin",
-    "MinecraftClientMixin",
-    "InstinctBarMixin",
-    "PlayerEntityAnimOverrideMixin",
-    "GameRendererMixin",
     "CursedMoonWorldMixin",
-    "MoonPhaseRenderMixin",
-    "CursedMoonLightmapMixin",
     "ItemStackMixin",
     "LivingEntityDeathMixin",
     "PlayerEntityPoseMixin",
     "CustomEdibleItemMixin",
     "CancelEntityStepSoundMixin",
-    "AdjustItemHoldFeatureRendererMixin",
-    "AdjustItemHoldPlayerRendererMixin",
-    "IEntityRenderDispatcherAccessor",
-    "EntityRenderDispatcherMixin",
     "EntityLootingMixin",
-    "FeralPlayerEntityRendererMixin",
-    "ElytraEntityModelMixin",
-    "PlayerEntityRendererFallFlyingMixin",
     "accessor.ActiveTargetGoalAccessor",
     "mob.WitchEntityMixin",
     "mob.VexEntityMixin",
@@ -35,6 +20,23 @@
     "mob.RavagerEntityMixin",
     "mob.VindicatorEntityMixin",
     "block.SweetBerryBushBlockMixin"
+  ],
+  "client": [
+    "IEntityRenderDispatcherAccessor",
+    "CursedMoonLightmapMixin",
+    "MoonPhaseRenderMixin",
+    "PlayerEntityAnimOverrideMixin",
+    "InstinctBarMixin",
+    "OverrideSkinFirstPersonMixin",
+    "OverrideSkinThirdPersonMixin",
+    "PlayerEntityRendererFallFlyingMixin",
+    "GameRendererMixin",
+    "AdjustItemHoldPlayerRendererMixin",
+    "EntityRenderDispatcherMixin",
+    "FeralPlayerEntityRendererMixin",
+    "MinecraftClientMixin",
+    "ElytraEntityModelMixin",
+    "AdjustItemHoldFeatureRendererMixin"
   ],
   "injectors": {
     "defaultRequire": 1
